20.02.23

+++ src/main/java/pro/sky/simplerecipes/services/RecipeService.java
@@ -6,5 +6,5 @@ public interface RecipeService {
 
     Recipe addRecipe(Recipe recipe);
 
-    Recipe getRecipe(Integer id);
+    Recipe getRecipe(long id);
 }
